Displays result of calculations

+++ src/mainwindow/mainwindow.h
@@ -21,32 +21,32 @@ public:
 private:
     Ui::MainWindow *ui;
 
-    static long double fun1(long double x, long double y) {
+    static double fun1(double x, double y) {
         return sin(y + 0.5) - x - 1;
     }
 
-    static long double fun2(long double x, long double y) {
+    static double fun2(double x, double y) {
         return y + cos(x - 2);
     }
 
-    static long double der1_x(long double x, long double y) {
+    static double der1_x(double x, double y) {
         return -1;
     }
 
-    static long double der1_y(long double x, long double y) {
+    static double der1_y(double x, double y) {
         return cos(y + 0.5);
     }
 
-    static long double der2_x(long double x, long double y) {
+    static double der2_x(double x, double y) {
         return -sin(x - 2);
     }
 
-    static long double der2_y(long double x, long double y) {
+    static double der2_y(double x, double y) {
         return 1;
     }
 
-    static void inverse_matrix(long double matrix[2][2]) {
-        long double det, element;
+    static void inverse_matrix(double matrix[2][2]) {
+        double det, element;
         det = matrix[0][0] * matrix[1][1] - matrix[0][1] * matrix[1][0];
         element = matrix[1][1];
         matrix[1][1] = matrix[0][0] / det;
